% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_loadings}
\alias{get_loadings}
\title{Get loadings}
\usage{
get_loadings(trainobj, what = c("finalModel", "CV"), xykeep = c("x", "y",
  "both"), rep = NA, remove_empty = TRUE)
}
\arguments{
\item{trainobj}{The \code{train} object from \code{caret}, fit using mixOmics spls.}

\item{what}{Which model to return loadings from. Either \code{"finalModel"} or \code{"CV"}.}

\item{xykeep}{Return loadings from x or y? Either \code{"x"}, \code{"y"}, or \code{"both"}.}

\item{rep}{If \code{what = "CV"}, and trainobj was fit using repeated cross-validation, choose a repeat (e.g. "Rep1") or leave as NA for all repeats (default).}

\item{remove_empty}{Remove loadings with a value of zero from output.}
}
\value{
A \code{data.frame} with the variables
\itemize{
  \item \code{var} - the variable
  \item \code{comp} - the component
  \item \code{loading} - the loading value (median if across CV folds / reps)
  \item \code{sd} - the standard deviation of the loading (if across CV folds / reps, NA otherwise)
  \item \code{xy} - is it a loading on x or y?
}
}
\description{
Returns the loadings of a mixOmics spls train object.
Either from the final model fitted on the full training data, or the median and sd across CV folds from the best performing tuning parameters. If there are multiple CV repeats, these are separated or can be chosen.
}
\examples{
library(caret)
x <- data.frame(matrix(rnorm(1000),nrow = 100))
y <- rnorm(100)
PLS <- train(x = x, y = y, method = get_mixOmics_spls())
get_loadings(PLS)

get_loadings(PLS, "CV") \%>\%
  ggplot(aes(var, loading, ymin = loading - sd, ymax = loading + sd)) +
    facet_wrap(~ comp, scales = "free_x") +
    geom_errorbar() +
    geom_bar(stat = "identity")
}
