% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_objects.R
\name{get_mixOmics_sPLS}
\alias{get_mixOmics_sPLS}
\title{Returns a list object which caret uses to interface with mixOmics.}
\usage{
get_mixOmics_sPLS(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
A \code{list} object that can be used with \code{caret::train()}.
}
\description{
Returns a list object which caret uses to interface with mixOmics.
}
\examples{
library(caret)
x <- data.frame(matrix(rnorm(1000),nrow = 100))
y <- rnorm(100)
train(x = x, y = y, method = get_mixOmics_sPLS())
}
