% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{permutation_spls}
\alias{permutation_spls}
\title{Perform permutation test}
\usage{
permutation_spls(fit, permutations = 1000, tester = cor, ...,
  comp = "all", seed = as.integer(Sys.time()), strata = NULL,
  fitfunc = mixOmics::spls)
}
\arguments{
\item{fit}{A model from \code{mixOmics::spls} or a \code{train} object from \code{caret}, fit using mixOmics spls.}

\item{permutations}{The number of permutations to perform.}

\item{tester}{A function used to compare the X and Y components, such as \code{cor}, or a character variable with one of the built-in methods ("cor", "rsq", "p", or "rmse"). Custom functions must return symmetric matrix (like \code{cor})}

\item{...}{Additional arguments to \code{tester}, such as \code{method = "spearman"}.}

\item{comp}{The component(s) that we wish to compute p values for as a numeric vector (e.g. 2:4), or "all" (default: "all").}

\item{seed}{Random seed.}

\item{strata}{Value to constrain permutations.}

\item{fitfunc}{The function used to compute the permuted fits, currently only works with \code{mixOmics::spls}.}
}
\value{
A \code{list} with the elements
\itemize{
  \item \code{p.values} - the computed p values per component
  \item \code{true} - the true values (from \code{tester}) that will be compared to a null distribution
  \item \code{track} - the permutation-distribution of values (from \code{tester}) that the true value was compared against.
  \item \code{seed} - the random seed used, for reproducibility.
}
}
\description{
Performs a permutation test of an spls fit or a train object. Uses hyperparameters from the fit (or optimal  \code{caret::train}-fit), which may be anti-conservative (overfit to data) so beware.
}
\examples{
library(caret)
x <- data.frame(matrix(rnorm(1000),nrow = 100))
y <- rnorm(100)
PLS <- train(x = x, y = y, method = get_mixOmics_spls())
perms <- permutation_spls(PLS)
}
